backend/requirements_frozen.txt
aiofiles==25.1.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrdict==2.0.1
babel==2.18.0
bce-python-sdk==0.9.60
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.0
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cssselect==1.4.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.2
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
gliner==0.2.13
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
modelscope==1.34.0
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnxruntime==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.76
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.7.3
paddlepaddle==2.6.2
paddlex==3.4.1
pandas==2.3.3
passlib==1.7.4
pdf2docx==0.5.8
pillow==12.1.0
premailer==3.10.0
prettytable==3.17.0
protobuf==3.20.2
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdfium2==5.3.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
rarfile==4.2
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==0.50.0
sympy==1.14.0
termcolor==3.3.0
thop==0.1.1.post2209072238
tifffile==2025.5.10
tokenizers==0.22.2
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
ultralytics==8.3.20
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.40.0
visualdl==2.5.3
watchfiles==1.1.1
wcwidth==0.5.3
websockets==16.0
Werkzeug==3.1.5
